+++ src/Statistics.h
@@ -81,7 +81,7 @@ public:
 	Statistics();
 	void setup();
 	void addSpeed(float speed);  // in 0,1km/h
-	void addDistance(uint16_t dist, ESummaryType type = SUM_ESP_TOUR);
+	void addDistance(uint32_t dist, ESummaryType type = SUM_ESP_TOUR);
 	void reset(ESummaryType type);
 	void updateDistance(uint32_t dist);
 	void addHR(int16_t heartrate);
